Change interior mutability of GraphicsWindow
The interior mutability was visible to the outside via new() returning a
Rc<RefCell<..>>.

When doing deep recursions during rendering for example, that window is
mutably borrowed. That in turn prevents us from adding further members
such as window scale properties, etc. that may be read *during* that
traversal as they'd require an immutable refernce to a already mutably
borrowed refcell contents.

So instead, make the individual fields refcells.