Use a different toml parser
Mysteriously, depending on toml_edit gives strange (reproducible) cargo
doc test errors due to "changed" dependencies. There must be some
"cross" dependency between this and the run-time library perhaps that
ends up with a recompilation somewhere. Switch to cargo-toml2 just for
reading Cargo.toml -- unfortunately the fixing doesn't work as it
reformats the entire fail (and errors out). But since most files are
changed this might be okay for now.