Fix build on Windows with warm cache
With a cold cache, neon-sys will be compiled as a dependency of the
nodejs buildings, which in turn will install node-gyp and run it to
build its bindings. On Windows, linkage against the nodejs import
library is needed for the correct symbol ordinals from nodejs.dll.
node-gyp takes care of downloading the header files and import library
into %HOMEPATH%\AppData\Local\node-gyp\Cache\<version>.

In a warm cache scenario, neon-sys is not rebuilt, node-gyp is not run
and the node.lib import library is not installed.

However neon-sys also propagates the path to node.lib (and the import
lib name) by parsing the node-gyp output, saving it to an environment
variable, exporting that via links=neon to neon-build and then
neon-build persists it into files its out directory. Finally we call
neon-build's setup() function, which reads those files and tries to link
sixtyfps-node-native also against node.lib. The absence of the import
library causes the build to fail.

This patch solves that by explicitly calling node-gyp to ensure the
files are installed.