Fix build with clang
Similar to the issue fixed in commit
de74ae4534d4b6fea6bf1729f9b6dbb5e086684e, clang appears to assume that
the float literal is a double our object literals and aggregate
initialization to float members (like the level of InkLevel) requires an
explicit cast.

I don't know of a way to determine the type of a struct field by index easily,
so this build fix instead uses C++17 destructuring to initialize the fields by index.

So

    InkLevel{ std::get<0>(o), std::get<1>(0) }
                              (^^ fails when get returns double and
                              field is float)

becomes

    InkLevel s; auto& [f1, f2] = s; f1 = std::get<0>(o); f2 = std:get<1>(o); return s;