Fix CheckBox height in the ugly style
With the previous grid layout implementation the checkbox minimum/maximum height of 40lx
was ignored in the todo app and instead it was
rendered using 20lx, which looks fine.

With the stretch layout the outer constraint of 20lx and the inner constraint of 40lx
create an impossible situation where eventually the
checkboxes in todo get misplaced.