Fix failing syntax tests
After commit 0c8a31211e3e8118b37d2c4d30d85385c6b21024 we're using an
older version of the regex crate, that uncovered a bug in our tests:

We were looking for

   Use parentheses to disambiguate && and ||

in the output messages by interpreting this as a regular expression. The
newer version of regex considered the || in the above regex as
alternations, empty alternations. So they basically always matched.

The old regexp that's pinned due to the above commit doesn't support
empty alternations, so the || need to be escaped. That uncovered a bug
in the regex itself, where it turns out that the word "between" was
missing after "disambiguate".