Fix hang on startup on some Linux systems
When the fontconfig dev package is not installed, servo-fontconfig will
fall back to building its own copy, which may be incompatible with
what's on the system. The result is a horrible start-up time and this
patch opts to instead require fontconfig to be available at build time
and error out otherwise.

Fixes issue #88