Avoid an unusable window due to invalid constraints
With manually placed maximum-width and minimum-width bindings it's
possible to create a layout where the minimum is greater than the maximum.

In the unlikely event of that happening, swap min/max so that the window
is resizeable. Otherwise at least on macOS the window manager tries to
apply both and the window funnily jumps.