Fix gallery layout
The layout before this patch resulted in constraints where the maximum
width was less than the calculated minimum width.
To avoid this situation, this patch rewrites the layout to use a central
vertical layout and places the items that we don't want to stretch into
horizontal boxes with a spacer.