Fix relative file imports from relative sources with the CLI viewer
Calling

    viewer ./examples/printerdemo/ui/printerdemo.60

would fail trying to resolve the relative path import "common.60" from
printerdemo.60, because we would apply the logic for resolving relative
include paths to be relative to the input file *also* for the initial
include path, derived from the main input file. That would result in a
search path

    ./examples/printerdemo/ui/./examples/printerdemo/ui

and that doesn't work :-)

So apply the resolution of relative include paths to be relative to the
input file only to those in the compiler configuration.