Fix panic during rendering with GL backend
During rendering the GLRenderer's Option<WindowedContext<NotCurrent>>
is None. That however leads to any font_metrics() calls such as due to
implicit sizing to panic, because the call requires access to the
window's scale factor.

This is fixed by keeping a shared reference to the PossiblyCurrent
WindowedContext also in the GLRenderer.