Fix regular key press event handling with winit
With commit

    9ca87ab3121a25813021fec9e63886ffc9adae2c

and

    c62b0e5316ac32a1f94e97d74a890251c2627aeb

the winit event loop sends our KeyPress for
winit::event::WindowEvent::KeyboardInput only when the virtual key code
matches our specially encoded keys. For regular keys we rely on
winit::event::WindowEvent::ReceivedCharacter, for which we sent only
KeyRelease (which TextInput listens for), but not KeyPressed. With
this patch we simulate press & release.